made custom parser for rho, added debugging prints, and added chems to variable display
CEA raw output was in exponential form only for rho and because of how it was printed I couldn't make a parser that worked for everything so I made rho its own. isp is not getting parsed correctly but it's less important at the moments so I left it as is.